[OLH-2293] force cloud formation update [skip canary] (#446)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->
[OLH-2293] force cloud formation update [skip canary]

### What changed
<!-- Describe the changes in detail - the "what"-->
Removed secret assign as it has not been refreshed. This will force an
update as tested in Dev. This PR will be reverted to then force the
correct secret value to be assigned.

### Why did it change

<!-- Describe the reason these changes were made - the "why" -->
Cloud formation is not updated secret value.

### Related links

<!-- List any related PRs -->
<!-- List any related ADRs or RFCs -->

## Testing

<!-- Provide a summary of any manual testing you've done, for example
deploying the branch to dev -->
Deployed to dev to force update secret value. 

[OLH-2293]:
https://govukverify.atlassian.net/browse/OLH-2293?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ