OLH-2303 - Trigger Pre-merge test to validate as part of Github Merge (#496)
## Proposed changes

OLH-2303 - Trigger Pre-merge test to validate as part of Github Merge
   
### What changed

 Make pre-merge a part of OLH release process by doing the following:
 
 Ensure merge to main is via a merge queue
As part of status check before merge, do a pre-merge test run. This
process deploys the merge candidate to a test environment and carries
out a series of end to end test, if all passes, proceeds with the merge
Define new trigger and filter to ensure only run in mere queue
Add ability to skip pre-merge test by including skipPreMerge in last 3
commits

### Why did it change

Make pre-merge testing a part of OLH release process. 

### Related links

## Checklists

### Environment variables or secrets

- [ ] No environment variables or secrets were added or changed

### Permissions

## Testing
Created a branch with similar configurations as the main branch, created
a pr against that branch and demonstrated a successful pre-merge test
run when skipPreMerge text is not in the last 3 commits messages and
where it is, demonstrated that the pre-merge test is skipped.

## How to review



## Additional Steps prior to merging
Ensure main branch has the same config as below
 
<img width="284" alt="image"
src="https://github.com/user-attachments/assets/c39ea7bb-4279-4f84-9151-df31585a6794"
/>