OLH-2399-stop-tx-ma-sending-us-the-wrong-data (#524)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->
[OLH-2399] Drop HMRC events & Fix last seen pretty

### What changed
<!-- Describe the changes in detail - the "what"-->
Dropped HMRC event's as they are not OLH users.
Last seen pretty isn't meant to be used but is still available in the
database. Updating to the correct value so the table will fix itself
over time.

### Why did it change
<!-- Describe the reason these changes were made - the "why" -->
HMRC are sending event's to be processed when they shouldn't be. This
has filled up our user_services table with 5m records that shouldn't be
there.
Last seen pretty date was expected in milliseconds turned out to be
seconds in production and was too late to fix.

## Checklists

<!-- Merging this PR deploys to production. Please answer accurately.
-->

### Environment variables or secrets

- [x] No environment variables or secrets were added or changed

## Testing

<!-- Provide a summary of any manual testing you've done, for example
deploying the branch to dev -->
## How to review

<!-- Describe any non-standard steps to review this work, or highlight
any areas that you'd like the reviewer's opinion on -->
Deployed to Dev and validated that hmrc event's were not processed by
pushing event's were different client_ids to the sns user_services
format queue.

[OLH-2399]:
https://govukverify.atlassian.net/browse/OLH-2399?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ