OLH-2404-remove-hmrc-events-from-user-services-table (#525)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->
[OLH-2404] Remove HMRC event's from user services table

### What changed
<!-- Describe the changes in detail - the "what"-->
Created a lambda that will run on test to remove all data from the
user_services table where the user_id does not start with urn

### Why did it change
<!-- Describe the reason these changes were made - the "why" -->
Officially all user_ids are meant to start with urn
HMRC have been sending us event's that we should not have processed. 
These user_ids do not start with urn and are not one login users.
They did not need to be stored in our user_services table. 

### Related links

<!-- List any related PRs -->
<!-- List any related ADRs or RFCs -->

## Checklists

<!-- Merging this PR deploys to production. Please answer accurately.
-->

### Environment variables or secrets

- [x] No environment variables or secrets were added or changed

<!-- Delete if changes DO NOT include new environment variables or
secrets -->

- [x] Added parameters to Cloudformation template
- [ ] Added new secret or systems manager parameter
- [ ] Added new environment variable

### Permissions

- [x] This PR adds or changes permissions

## Testing

<!-- Provide a summary of any manual testing you've done, for example
deploying the branch to dev -->
Deployed to Dev and ran. Validated that all item's in the user_services
table where the user_id != urn... were deleted.

## How to review

<!-- Describe any non-standard steps to review this work, or highlight
any areas that you'd like the reviewer's opinion on -->


[OLH-2404]:
https://govukverify.atlassian.net/browse/OLH-2404?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ