removed provisioned concurrency (#562)
## Proposed changes

<!-- Provide a general summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX] PR Title` -->
BAU - Performance Optimisations

### What changed

<!-- Describe the changes in detail - the "what"-->
Previously provisioned concurrency was added to facilitate the NFR which
had a high number of users per second. This NFR was set too high and did
not take into account real world traffic. Whilst this NFR is being
adjusted, we are reverting our configuration to save costs.
Additionally, provisioned concurrency is not required for the backend as
we are aiming for eventual consistency. This means lambda cold start
times do not matter and the queues will ensure they are completed.

### Why did it change

<!-- Describe the reason these changes were made - the "why" -->
Save costs ~$300 per month

### Testing

<!-- When working with feature flags, features that are flagged off
should not be made available in production -->
<!-- Delete if changes do NOT include any feature flags -->
Deployed to Dev and ran a performance test against it