BAU: Unpin cfn-lint
In #580 we pinned to an old version of cfn-lint to avoid a bug. That
bug's been fixed (see https://github.com/aws-cloudformation/cfn-lint/pull/4041)
so we can unpin and get the latest version.

I did consider leaving the pin in place to avoid these sorts of errors
in the future, but it'd add another dependency upgrade that we'd have to
look for, test and upgrade manually and I vaguely remember a
conversation from a while back where we decided that that wasn't worth
the tradeoff.