OLH-2521: Mock the RP registry library in tests
Now the RP data is managed in a separate library we can update the data
without touching the library. We merged a change to the GOV.UK email
subscriptions client data in that library, but that RP is used as the
example data in the test suite here.

We don't want to end up with a change that's merged in one place that
has the potential to break tests elsewhere. Instead, mock the function
call in this app as we would for any other external library.