Bau refactor hmrc data delete script (#605)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->
[BAU] Refactor HMRC data delete script to continue from the last
evaluated key

### What changed
<!-- Describe the changes in detail - the "what"-->
enabled passing in the last evaluated key as a env var

### Why did it change
<!-- Describe the reason these changes were made - the "why" -->
In production there are too many records in the activity log table for
it to be processed within the 15 min lambda limit.
This change will allow us to continue from the last completed instead of
iterating through all the records that have already been checked.

## Testing

<!-- Provide a summary of any manual testing you've done, for example
deploying the branch to dev -->
Deployed to dev and tested.