Olh 2294 backend lambda event triggers are not pointing to live alias (#617)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->
[OLH-2294] Point SNS topics and Permissions to alias live version of
backend Lambda's

### What changed

<!-- Describe the changes in detail - the "what"-->
Pointed SNS topics and permissions to live version of lambda's.

### Why did it change
<!-- Describe the reason these changes were made - the "why" -->
We introduced the live alias following the implementation of canary
deployments. However, some of the event triggers were still pointing to
the function instead of the live alias. This meant that during canary
deployment, the wrong lambda versions were being tested for some
lambda's.
### Permissions

- [ ] This PR adds or changes permissions

## Testing

<!-- Provide a summary of any manual testing you've done, for example
deploying the branch to dev -->
Deployed to dev and validated.

[OLH-2294]:
https://govukverify.atlassian.net/browse/OLH-2294?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ