BAU: Increase timeout of delete activity log lambda
This lambda seems to timeout once every couple of weeks. When we retry
the event it always succeeds. It does seem to sometimes take a long time
to delete items from the table (users can have many entries) so double
the timeout to 10s to hopefully stop our alarms going off.