OLH-2775: Remove start and finish processing logs (#661)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->

### What changed

Remove start and finish processing logs from all the Lambdas.

### Why did it change

We're getting close to breaching the GDS-wide Splunk data quota and
teams have been asked to check their logging and reduce where possible.

Our backend emits about 10 million log events per day and these
processing messages are the vast majority of that.

We initially added them to help with debugging or tracing a message
through the system. In practice we've never used the message IDs in the
~3 years that these data pipelines have been running. Failed events are
put on the DLQs anyway, so if we do need to see the IDs we can always
pull the messages from the DLQ and see the IDs. The built in Lambda logs
also contain a request ID to look at all logs in that execution.

I've left all the error messages in place so we're not losing any
information when we do need to debug a problem.

### Related links

https://gds.slack.com/archives/C049ASDACJV/p1749715968265739

## Checklists

<!-- Merging this PR deploys to production. Please answer accurately.
-->

### Environment variables or secrets

- [x] No environment variables or secrets were added or changed