OLH-3061: hardcode Dynatrace layer ARN (#713)
### What changed

Hardcodes the Dynatrace layer ARN. This is not the latest ARN which will
come in a subsequent PR.

### Why did it change

There’s a bug in the `{{resolve:secretsmanager}}` syntax that prevents
the current configuration from deploying the latest available Dynatrace
layer. The fix is to just hardcode the layer ARN and manually keep it up
to date.

Co-authored-by: Michael Henson <12057877+liegeandlief@users.noreply.github.com>