OLH-3171: Create a new lambda that consumes SQS email events and sends an email via Gov.Notify (#734)
### What changed

Adds a Lambda function which consumes messages from the notification
queue and sends Notify emails.

Currently it is only configured to send global logout emails but it can
be easily expanded to send other emails by adding to the
`NotificationType` enum, the `messageSchema` and the `NOTIFYTEMPLATEIDS`
environment variable.

The Lambda implements partial batch processing in order to be able to
handle batches of messages whilst being able to return only failed
messages to the queue to be retried.

These changes are currently deployed to dev and can be tested by going
to SQS in the dev AWS console and sending messages on the
`account-mgmt-backend-NotificationQueue` queue. Some example messages
which can be used to test it (you'll need to put your email address in
the message):

```
{"notificationType":"GLOBAL_LOGOUT","emailAddress":"YOUR_EMAIL_HERE","loggedOutAt":"2023-01-01T14:00:00.000Z","ipAddress":"192.168.1.3","userAgent":"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/140.0.0.0 Safari/537.36","countryCode":"FR"}
```

```
{"notificationType":"GLOBAL_LOGOUT","emailAddress":"YOUR_EMAIL_HERE","loggedOutAt":"2023-01-01T14:00:00.000Z","ipAddress":"192.168.1.3","userAgent":"Not a valid user agent","countryCode":"FR"}
```

```
{"notificationType":"GLOBAL_LOGOUT","emailAddress":"YOUR_EMAIL_HERE","loggedOutAt":"2023-01-01T14:00:00.000Z"}
```

---------

Co-authored-by: Michael Henson <12057877+liegeandlief@users.noreply.github.com>