BAU: Log the item size when writing to DynamoDB
In this table we have 1 item per user, and as they log in to more
government services we append entries to a list attribute on the item.

DynamoDB has a per-item size limit of 400kb and at the moment we have
no monitoring for how big the items are in the table. As we're
constantly appending data it's possible that the item for a high-volume
user could be approaching 400kb.

Log out the item size before we write to the table. This is the best
approximation of the size of the item represented as JSON. DynamoDB
won't store the item as JSON but if this number is approaching 400kb
then we'll at least know we need to take action.