BAU: Log the item size when writing to DynamoDB (#792)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->

### What changed

Log the item size when writing user services to DynamoDB. This is the
best approximation of the size of the item represented as JSON. DynamoDB
won't store the item as JSON but if this number is approaching 400kb
then we'll at least know we need to take action.

Once this is merged in I'll set up some monitoring in Splunk / Dyntrace
to see what the results are.

### Why did it change

In this table we have 1 item per user, and as they log in to more
government services we append entries to a list attribute on the item.

DynamoDB has a per-item size limit of 400kb and at the moment we have no
monitoring for how big the items are in the table. As we're constantly
appending data it's possible that the item for a high-volume user could
be approaching 400kb.

## Checklists

<!-- Merging this PR deploys to production. Please answer accurately.
-->

### Environment variables or secrets

- [x] No environment variables or secrets were added or changed

## Testing

<!-- Provide a summary of any manual testing you've done, for example
deploying the branch to dev -->

## How to review

<!-- Describe any non-standard steps to review this work, or highlight
any areas that you'd like the reviewer's opinion on -->