OLH-3229: Log the error message if the API call fails
We recently saw an outage of ~10 minutes for the API this Lambda calls
which meant 11 requests failed. We throw an error in this case, so they
ended up on the DLQ, but there was no messages in our application logs
to explain what had happened.

Log out the error message before throwing so we have an easier time
debugging next time there's a problem.

The main logic function in this Lambda had no test coverage at all, so
I've also added a set of unit tests for the logging and all the
existing behaviour.