OLH-3229: Log the error message if the API call to GOV.UK fails (#795)
## Proposed changes

<!-- Provide a summary of your changes in the title above -->
<!-- Include the Jira ticket number in square brackets as prefix, eg
`[OLH-XXXX]: PR Title` -->

### What changed

Log the error message if the API call to GOV.UK fails.

The main logic function in this Lambda had no test coverage at all, so
I've also added a set of unit tests for the logging and all the existing
behaviour.

### Why did it change

We recently saw an outage of ~10 minutes for the API this Lambda calls
which meant 11 requests failed. We throw an error in this case, so they
ended up on the DLQ, but there were no messages in our application logs
to explain what had happened.

Log out the error message before throwing so we have an easier time
debugging next time there's a problem.


## Checklists

<!-- Merging this PR deploys to production. Please answer accurately.
-->

### Environment variables or secrets

- [x] No environment variables or secrets were added or changed